+++ src/views/home/right/OutputRight.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { ResultFull } from '../../services/voting/SuperElection';
+import { ResultFull } from '../../../services/voting/SuperElection';
 import ContingencyBlock from './blocks/ContingencyBlock';
 import { MemoizedIRVBlock } from './blocks/IRVBlock';
 import { MemoizedStaticBlock as StaticBlock } from './blocks/StaticBlock';
